Refactor: страницу Vacancy page
